feat(angular): use the context for getting the repository and host urls (#217)
* feat(angular): use the context for getting the repository and host urls

This will result in more flexibility since we now use the logic already existing in conventional-changelog-core and at the same time make the code simpler. The logic now matches the one already present in the templates for inserting a link to the commit.

Add test verifying we get the right url from the context.